# specs/openapi.yaml
openapi: 3.0.3
info:
  title: Приложение для федерации тяжелой атлетики г. Москвы (ФТАМ)
  description: Приложение для спортcменов, тренеров, организаторов и представителей ФТАМ - автоматизация деятельности ФТАМ.
  version: 1.0.11

# servers:
#   - url: https://wlmoscow
...
paths:
...
  /competitions/{id}/results:
    get:
      tags:
        - competitions
      summary: Получить все результаты соревнования
      operationId: getResultsByCompId
      parameters:
        - name: id
          in: path
          description: ID соревнований
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllCompResults'          
        '400':
          description: Invalid ID supplied
        '500':
          description: Internal Server Error
...
  /competitions/{id}:
    delete:
      tags:
        - competitions
      summary: Удалить соревнование
      operationId: deleteComp
      parameters:
        - name: id
          in: path
          description: ID соревнования
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
            description: Successful operation
        '401':
          description: Unauthorized
        '404':
            description: Not found
        '500':
            description: Internal Server Error
...
  /competitions:
    get:
      tags:
        - competitions
      summary: Получить все соревнования
      operationId: getComps
      parameters:
        - name: name
          in: query
          description: Название мероприятия
          required: false
          schema:
            type: string
            example: "Кубок Москвы"
        - name: date
          in: query
          description: Дата начала (возможно только год)
          required: false
          schema:
            type: string
            format: datetime(utc)
        - name: city
          in: query
          description: Город
          required: false
          schema:
            example: "Москва"
            type: string
        - name: age
          in: query
          description: Возрастная группа
          required: false
          schema:
            example: "Юниоры, юниорки"
            type: string
        - name: sort
          in: query
          description: Сортировка. Список имен атрибутов, разделенных запятыми, для сортировки. Если перед именем атрибута стоит знак минус, то сортировка производится по убыванию, в противном случае — по возрастанию.
          required: false
          schema:
            example: date,-city
            type: string
        - name: page
          in: query
          description: Пагинация - номер страницы.
          required: false
          schema:
            example: 3
            type: integer
        - name: batch
          in: query
          description: Пагинация - количество элементов на странице
          required: false
          schema:
            example: 10
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllComps'
        '500':
          description: Internal Server Error
    post:
      tags:
        - competitions
      summary: Организация соревнований
      operationId: createComp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComp'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comp'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
  /sportsmen/{id}:
    get:
      tags:
        - sportsmen
      summary: Получить спортсмена по ID
      operationId: getSportsmanById
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsmanExt'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Sportsman not found
        '500':
          description: Internal Server Error
    patch:
      tags:
        - sportsmen
      summary: Обновление секретарем информации о спортсмене
      description: Обновление информации о сертификате и допуске на соревнования
      operationId: UpdateSportsman
      parameters:
          - name: id
            in: path
            description: ID спортсмена
            required: true
            schema:
              type: string
              format: uuid
      requestBody:
        description: Обновление информации о существующем спортсмене
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSportsman'
        required: false
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SportsmanExt'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Sportsman not found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - sportsmen
      summary: Удалить спортсмена
      operationId: deleteSportsman
      parameters:
        - name: id
          in: path
          description: ID спортсмена
          required: true
          schema:
              type: string
              format: uuid
      responses:
        '204':
            description: Successful operation
        '401':
          description: Unauthorized
        '404':
            description: Sportsman not found
        '500':
            description: Internal Server Error
...
  /competitions/{id}/sportsman:
    post:
      tags:
        - competitions
      summary: Запись спортсмена на соревнования
      parameters:
        - name: id
          in: path
          description: ID соревнования
          required: true
          schema:
              type: string
              format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompApplication'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompApplication'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: Not found
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    Comp:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        name:
          type: string
          example: Кубок Москвы
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        age:
          type: string
          example: Юниоры, юниорки
        min_sports_cat:
          type: string
          example: КМС
        antidoping:
          type: boolean
          example: true
...
    SportsmanExt:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        surname:
          type: string
          example: Иванов
        name:
          type: string
          example: Иван
        patronymic:
          type: string
          example: Иванович
        birthday:
          type: string
          format: date-time
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        gender:
          type: boolean
          example: false
        adoping_validity:
          type: string
          format: date-time
        access_validity:
          type: string
          format: date-time
        access_institution:
          type: string
          example: 1-й диспансер
...
    CompApplication:
      type: object
      properties:
        comp_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        weight_category:
          type: string
          example: 109+
        start_snatch:
          type: integer
          example: 100
        start_clean_and_jerk:
          type: integer
          example: 140
    CreateCompApplication:
      type: object
      properties:
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        weight_category:
          type: string
          example: 109+
        start_snatch:
          type: integer
          example: 100
        start_clean_and_jerk:
          type: integer
          example: 140
...
    CreateComp:
      type: object
      properties:
        name:
          type: string
          example: Кубок Москвы
        city:
          type: string
          example: Москва
        address:
          type: string
          example: Московская 1
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        age:
          type: string
          example: Юниоры, юниорки
        min_sports_category:
          type: string
          example: КМС
        antidoping:
          type: boolean
          example: true
...
    UpdateSportsman:
      type: object
      properties:
        moscow_team:
          type: boolean
          example: true
        sports_category:
          type: string
          example: КМС
        access_institution:
          type: string
          example: 1-й диспансер
        access_validity:
          type: string
          format: date-time
        antidoping_validity:
          type: string
          format: date-time
...
    AllComps:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/Comp'
...
    CompResult:
      type: object
      properties:
        sm_id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440001
        comp_name:
          type: string
          example: Кубок Москвы
        sm_fullname:
          type: string
          example: Иванов Иван Иванович
        beg_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        city:
          type: string
          example: Москва
        weight_category:
          type: string
          example: 109+
        snatch:
          type: integer
          example: 120
        clean_and_jerk:
          type: integer
          example: 200
        place:
          type: integer
          example: 2
    AllCompResults:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CompResult'
...